Refactoring: Separated model class, renamed Methods, separated CognitiveServicesCallHelper.
