+++ IdentifyService/IdentifyWeb/IdentifyWeb.csproj
@@ -149,7 +149,9 @@
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
-    <Compile Include="MultipartFormdataStreamBlobUploader.cs" />
+    <Compile Include="Models\CommonModels.cs" />
+    <Compile Include="Utility\CognitiveServicesCallHelper.cs" />
+    <Compile Include="Utility\MultipartFormdataStreamBlobUploader.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -166,7 +168,6 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="App_Data\" />
-    <Folder Include="Models\" />
   </ItemGroup>
   <PropertyGroup>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
